% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{test_function}
\alias{test_function}
\title{Test function for educational purposes}
\usage{
test_function(num1, num2)
}
\arguments{
\item{num1}{First number}

\item{num2}{Second number}
}
\value{
A sum of num1 + num2 squared
}
\description{
Test function for educational purposes
}
